+++ src/test/java/com/hotels/beans/transformer/MutableObjectTransformationTest.java
@@ -36,6 +36,7 @@ import com.hotels.beans.sample.mutable.MutableToFoo;
 import com.hotels.beans.sample.mutable.MutableToFooInvalid;
 import com.hotels.beans.sample.mutable.MutableToFooNotExistingFields;
 import com.hotels.beans.sample.mutable.MutableToFooSimpleNoSetters;
+import com.hotels.beans.sample.mutable.MutableToFooSubClass;
 
 /**
  * Unit test for all {@link Transformer} functions related to Mutable type Java Beans.
@@ -151,5 +152,20 @@ public class MutableObjectTransformationTest extends AbstractTransformerTest {
 
         //THEN
         assertThat(mutableObjectBean, hasProperty(AGE_FIELD_NAME, equalTo(AGE)));
+    }
+
+    /**
+     * Test transformation on an existing bean is correctly copied.
+     */
+    @Test
+    public void testTransformationOnAnExistingDestinationWorksProperly() {
+        //GIVEN
+        MutableToFooSubClass mutableToFoo = new MutableToFooSubClass();
+
+        //WHEN
+        underTest.transform(fromFooSubClass, mutableToFoo);
+
+        //THEN
+        assertThat(mutableToFoo, sameBeanAs(fromFooSubClass));
     }
 }